pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eduardo Beckhauser", email = ""},
]
dependencies = [
    "django>=4.2.6",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.1",
    "django-cloudinary-storage>=0.3.0",
    "python-dotenv>=1.0.0",
    "netifaces>=0.11.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"